+++ bittensor/_synapse/text_prompting/example.py
@@ -16,69 +16,64 @@
 # DEALINGS IN THE SOFTWARE.
 
 import torch
+import argparse
 import bittensor
+from typing import Union
 from typing import List, Dict
-bittensor.logging(debug=True)
-# import openai
-
-
+bittensor.logging( debug = True )
 
 # Create a synapse that returns zeros.
-class Synapse(bittensor.TextPromptingSynapse):
-    def _priority(self, forward_call: "bittensor.TextPromptingForwardCall") -> float:
-        return 0.0
+class Synapse( bittensor.TextPromptingSynapse ):
 
-    def _blacklist(self, forward_call: "bittensor.TextPromptingForwardCall") -> bool:
-        return False
+    def __init__( self, config: "bittensor.Config" = None ):
+        super().__init__( config )
 
-    def backward( self, messages: List[Dict[str, str]], response: str, rewards: torch.FloatTensor ) -> str:
-        # Apply PPO.
+    @classmethod
+    def add_args( cls, parser: argparse.ArgumentParser ):
         pass
 
-    def forward(self, messages: List[Dict[str, str]]) -> str:
+    @classmethod
+    def check_config( cls, config: 'bittensor.Config' ):
+        pass
+
+    def priority( self, request: Union[ bittensor.ForwardTextPromptingRequest, bittensor.BackwardTextPromptingRequest ] ) -> float:
+        return 0.0
+
+    def blacklist( self, request: Union[ bittensor.ForwardTextPromptingRequest, bittensor.BackwardTextPromptingRequest ] ) -> bool:
+        return False
+
+    def backward( self, messages: List[Dict[str, str]], response: str, rewards: torch.FloatTensor ):
+        pass
+
+    def forward( self, messages: List[Dict[str, str]] ) -> str:
         return "hello im a chat bot."
+    
 
-# Create a mock wallet.
+config = Synapse.config()
+config.axon.port = 9090
+config.axon.ip = "127.0.0.1"
+print ( config )
+syn = Synapse( config = config )
+print ( config )
+
+
 wallet = bittensor.wallet().create_if_non_existent()
-
-# Create a local endpoint receptor grpc connection.
 local_endpoint = bittensor.endpoint(
-    version=bittensor.__version_as_int__,
-    uid=0,
-    ip="127.0.0.1",
-    ip_type=4,
-    port=9090,
-    hotkey=wallet.hotkey.ss58_address,
-    coldkey=wallet.coldkeypub.ss58_address,
-    modality=0,
+    version = bittensor.__version_as_int__,
+    uid = 0,
+    ip = "127.0.0.1",
+    ip_type = 4,
+    port = 9090,
+    hotkey = wallet.hotkey.ss58_address,
+    coldkey = wallet.coldkeypub.ss58_address,
+    modality = 0,
 )
-
-metagraph = None # Allow offline testing with unregistered keys.
-axon = bittensor.axon(wallet=wallet, port=9090, ip="127.0.0.1", metagraph=metagraph)
-
-synapse = Synapse()
-axon.attach(synapse=synapse)
-axon.start()
-
-batch_size = 4
-sequence_length = 32
-# Create a text_prompting module and call it.
 module = bittensor.text_prompting( endpoint = local_endpoint, wallet = wallet )
 forward_response = module.forward(
     roles = ['user', 'assistant'],
     messages = [{ "user": "Human", "content": "hello"}],
     timeout = 1e6
 )
-backward_response = module.backward(
-    roles = ['user', 'assistant'],
-    messages = [{ "user": "Human", "content": "hello"}],
-    response = forward_response.response,
-    rewards = [1,2,3,4,5],
-    timeout = 1e6
-)
 
 
-# # Delete objects.
-del axon
-del synapse
-del module
+
